Split training register from ai.onnx register file
